+++ package.json
@@ -391,8 +391,8 @@
   "scripts": {
     "build": "yarn build:cljs && yarn build:js",
     "build-embedding-sdk": "yarn && yarn build-release:cljs && yarn embedding-sdk:generate-package && yarn build-release:sdk && yarn build-embedding-sdk-cli && yarn embedding-sdk:tsc && yarn embedding-sdk:fixup-types-imports && yarn embedding-sdk:generate-nextjs-compat",
-    "build-embedding-sdk:pure": "WEBPACK_BUNDLE=development node tsup.embedding-sdk.config.js",
-    "build-embedding-sdk:watch": "yarn embedding-sdk:generate-package && yarn embedding-sdk:generate-nextjs-compat && WEBPACK_BUNDLE=development node tsup.embedding-sdk.config.js --watch",
+    "build-embedding-sdk:pure": "WEBPACK_BUNDLE=development node tsup.embedding-sdk.config.mjs",
+    "build-embedding-sdk:watch": "yarn embedding-sdk:generate-package && yarn embedding-sdk:generate-nextjs-compat && WEBPACK_BUNDLE=development node tsup.embedding-sdk.config.mjs --watch",
     "build-embedding-sdk-cli": "rspack --config rspack.embedding-sdk-cli.config.js",
     "build-embedding-sdk-cli:watch": "yarn build-embedding-sdk-cli --watch",
     "build-hot": "rm -f target/cljs_dev/cljs.core.js && yarn build:cljs && yarn build-hot:js",
@@ -404,7 +404,7 @@
     "build-release": "yarn build-release:cljs && yarn build-release:js",
     "build-release:cljs": "yarn clean-release:cljs && shadow-cljs release app",
     "build-release:js": "WEBPACK_BUNDLE=production rspack build",
-    "build-release:sdk": "WEBPACK_BUNDLE=production node tsup.embedding-sdk.config.js",
+    "build-release:sdk": "WEBPACK_BUNDLE=production node tsup.embedding-sdk.config.mjs",
     "build-release:static-viz": "WEBPACK_BUNDLE=production rspack --config rspack.static-viz.config.js",
     "build-shared": "yarn && webpack --config webpack.shared.config.js",
     "build-static-viz": "yarn && yarn build-release:cljs && WEBPACK_BUNDLE=production rspack --config rspack.static-viz.config.js",
